fix config envars and rename stuff
