+++ src/config/reddit_config.py
@@ -13,9 +13,9 @@ AUTH = {
 }
 
 for config in list(AUTH):
-  config_envar = f"REDD_BOT_{config}".upper()
+  config_envar = f"REDDIT_BOT_{config}".upper()
   if os.environ.get(config_envar):
-    # print(os.environ.get(f"BOT_{config}".upper()))
+    log.info(f"loading {config_envar} from envar")
     AUTH[config]=os.environ.get(config_envar)
   else:
     log.info(f"no environment config for: {config_envar}")
